Record button, displays real-time audio waveform
Added a "record" button to the GUI with a connected action that opens a PyAudio stream using the user-selected input device.

For a test period of currently 5 seconds (set arbitrarily), the audio input stream is plotted in real-time as a waveform using Matplotlib.

PyQtGraph will eventually be used for this purpose instead as its speed makes it better suited for real-time applications.