Added song playhead
A "playhead" widget now moves across the waveform plot of the audio file to represent song progress.

Currently, it is hidden until the user presses the play button for the first time as it starts in the incorrect place when the window is first rendered. A fix should be found eventually.