NotePredictor class now uses note onsets
The Librosa onset_detect function is now implemented in the NotePredictor's predict method to improve the accuracy of note detection.

The list of predictions is now shortened to only show pitch estimations of corresponding note onsets, to attempt to address the issue of held notes being counted multiple times due to constant timesteps.

Test file now displays a basic Matplotlib plot of the estimated notes to better visualise predictions and spot errors faster.