Rewrote AudioStreamHandler Class
Replaced PyAudio with sounddevice module in the AudioStreamHandler class, improving readability and lending itself better to audio processing due to NumPy support.

Replaced open_stream and close_stream methods with a Record method that takes an input device index as an argument and saves 5 seconds of input audio to a temporary WAV file. This will be expanded on to run continuously and use threading to perform real-time pitch detection on the user's guitar input.

Temporarily added a record button to the main GUI that calls this method for testing purposes.