Improvements to main.py, added count-in timer
Moved all widget styling to _set_styles method in MainWindow class.

Added a count-in timer when a song is played that is synchronised to its tempo and offset by its time position.

AudioLoadHandler now has a play_metronome method that uses sounddevice to play a metronome sound. The class's load function now also gets the loaded song's tempo and position of the first detected beat that is used to offset the timing of the initial metronome sound during a count-in.

Added methods to the MainWindow class to facilitate the use of a count-in timer.

WaveformPlot's private _draw_plot method has been made public and now takes an AudioLoadHandler as an argument to allow plotting of new audio data when a different song is loaded.