Added guitar track volume slider
The AudioLoadHandler class now loads the two separated guitar and no_guitar tracks of a song. The output stream now sums the amplitudes of the two tracks together to get the full mix. This allows the volume of the guitar in the mix to be adjusted by scaling the values (amplitudes) of its audio time series ndarray by a factor between 0 and 1.

A QSlider with a range between 0 and 100 has been added to the main GUI that allows a user to freely adjust the guitar track's volume, as well as labels indicating its current volume status.

Removed the test "record" button, and pressing the play button now starts the audio input's recording and processing loop.

TODO:

All playback logic in the AudioLoadHandler class should ideally be moved to a new AudioPlaybackHandler class.

An issue that needs to be addressed with the recording loop is that it seems to be blocking or freezing the GUI when recording is stopped (i.e. when the pause button is pressed).