Fix looping bugs, separate skipping & looping code
The audio output stream's callback method has been changed to fix a bug that occurred when the song's playhead was positioned after a looping section. The song is now allowed to continue playing while outside of a loop.

In the main file, code from the _waveform_pressed method was divided into two new methods to better follow the single responsibility principle.

The _loop_marker_set method performs the setting of a new time position for the left or right loop markers, and initiates song looping when both are set. In this method, another bug that occurred when a loop marker was set to 0 was addressed. As well, a minimum loop section time constraint of 2 seconds has been set for the time being to avoid other bugs resultant from too small a loop time.

The _skip_song_position method performs song seeking functionality.

Further small improvements to code readability in the main file have been made, such as making long strings easier to parse, and rewriting some comments.