Add loop toggle button to GUI
A loop toggle button has been added displayed to the right of the play/pause button in the playback controls.

The button's background colour is desaturated when looping is disabled to indicate the change to the user.

Updating a marker position by clicking the waveform will automatically re-enable looping for user convenience.