Add skip forward and back buttons
Skip forward and skip back buttons have been added to the playback controls that add or subtract five seconds from the current song time.

Conditionals were added to prevent a user from accidentally skipping past the end of loop/end of song, and skipping before start of loop/start of song.

An in_loop_bounds method has been added to the AudioPlayback that returns True if the song is currently looping and its position is within the bounds of the loop markers. This is a convenience function to avoid long conditionals.