Add compare_pitches test, small optimisation
Created a compare_pitches unit test that uses a fixture for mock user and song pitches, asserting that the correct score results are calculated.

A minor optimisation to the scoring algorithm was added, breaking the second loop early if the first song note's array is empty.

Second conditional changed to check second song note array is not empty, or the first index of both song notes are equal. This works thanks to short-circuiting behaviour.

Resultant score data from compare_pitches is now unpacked before being sent to the GUI for better readability.