Merge I/O streams into single streaming class
The individual input and output streams from the AudioInput and AudioOutput classes respectively have been replaced with a single sounddevice Stream object which handles I/O logic in a shared callback function. This has proven to be much more efficient than using separate streams.

To facilitate this change, the AudioInput and AudioOutput classes have been moved to a shared audio_streaming module, and a new AudioStreamHandler class has been created. This class runs the stream, as well as handling the playback controls due to their functionality relying on the stream callback function.

The scoring functions and the AudioInput class's _process_recording function have been moved to a new ScoringSystem class. As this class is not integrated into the application yet, this means the user scoring is currently temporarily disabled.

Adjusted the test_compare_pitches file to match these changes.