Preload Basic Pitch model for worker process
A PITCH_MODEL constant has been added to the config file that preloads the Basic Pitch model to avoid unnecessary and expensive loading times.

The max number of workers for the ScoringSystem's process pool executor has been set to 1, meaning only two cores at most should be utilised by the application. The created worker process now immediately preloads the Basic Pitch model rather than loading it when the process_recording function is called for the first time. This fixes the problem of noticeable delay the first time scores are calculated.