Re-implement overlapping windows for note scoring
The rec_buffer_size in the config file has been changed to a length of 2 seconds, and a new rec_overlap_window_size value has been added with a length of 1 second (44100 frames). An in_overlap attribute has been added to the AudioStreamHandler class to store audio data that will be appended to the buffer.

The audio input buffer now continuously queues and dequeues 1 second batches of recorded audio data, allowing the scoring system to properly account for notes at the edges of recordings which are currently being missed.

An issue resultant from this is that most of the user's note scores are now being counted twice. A temporary fix for this until a better solution can be found is to simply half the scores returned by the process_recording function.

Temporarily added a hard-coded additional offset of 150ms to user note times, as current method of using sounddevice to retrieve input device latency doesn't account for the round-trip delay of the guitar's FX chain, the audio interface's latency, etc.