Create Basic Pitch model tuning notebook
A notebook has been created that currently evaluates the pitch prediction model's performance on a test song against its community-tabbed guitar notes used as a ground truth.

The py_midicsv module is currently used to convert the guitar tab MIDI file to a CSV, which is then read as a pandas DataFrame.

Necessary preprocessing steps are then performed on the predictions and test notes, such as organising them into intervals and pitches ndarrays, and converting pitches from MIDI notes to Hz.

The mir_eval module is used to get metrics such as F1 score, precision, and recall for the model's prediction.