Use the IDMT-SMT-Guitar Dataset for tuning
A guitar dataset (IDMT-SMT-Guitar) with polyphonic and monophonic audio and XML-formatted annotations is now used for parameter tuning of the Basic Pitch model to optimise for guitar performance.

A train-test split of 80:20 is used, and a simplistic implementation of a randomised parameter search (inspired by sklearn's RandomizedSearchCV function) was created.

Initial hyperparameter tuning resulted in a model that achieved an F1 score of ~0.77 on the training set.