Add multiple windows functionality to GUI
All practice mode GUI code has been moved to its own PracticeWindow class in a separate module. A new SetupWindow class has been added that provides a button that allows the user to select a song to perform from their local files using the QFileDialog.getOpenFileName method.

Both windows are subclasses of PyQt's QWidget, allowing the MainWindow in the main file to switch between them using a QStackedWidget. The PracticeWindow is initialised when a signal is sent from the SetupWindow indicating a song has been selected.