Add performance time testing to scoring pipeline
Using time.perf_counter, the total elapsed time from receiving a new 1 second batch of audio to the output of the resultant processed score is now timed for testing purposes.

Added a zero_buffers function to the AudioStreamHandler class called when the song position is skipped to empty all current audio from the buffers.