Prepare codebase for PyInstaller
Changed paths in codebase for case in which the application is being run as an executable or a script.

Added a Directories section to the config file that contains the absolute paths of all parent directories needed while running the application.

Changed the preload module, performing the preloading logic in a preload function now called at the top of the main file that also updates the config file's Directories section paths based on if the application is being run as an executable.

Added code to the main function that closes the executable's splash screen when the GUI has loaded.