Improve GUI, add main.spec file for PyInstaller
The QSS stylesheet file has been changed to improve aesthetics of certain widgets, the main window is now resizeable with minimum dimensions of 1080x432.

A main.spec file has been created for PyInstaller when the application is bundled into a single-file executable.

The input and output latency config variables are now derived from the stream's latency values, which is more accurate.