Refactor code for PyInstaller one-dir executable
The preload function has been changed to preload_directories, and now saves the necessary directory as os.environ variables rather than in the config file.

Downloaded the htdemucs_6s model and storing it in a demucs_models folder to allow use of demucs locally. Added the folder as a "--repo"" option in the demucs separation args to make it look for the local model.

main.spec file changed for one-dir executable instead of one-file. Now additionally bundles the htdemucs_6s model, ffmpeg, and SDL3 for demucs and Torch functionality. As well, in the preload_directories function, the location of ffmpeg.exe has been added to the PATH variable to allow the executable to find it.

The config.ini file has been moved to the data directory for the executable.