Add new labels to GUI and improve layout
Added "previous accuracy" and "previous score" labels (currently non-functional), and a "swing" label that informs the user in milliseconds how much on average they are rushing or dragging by.

Changed the layout of the practice screen song information and increased the font size of more important info to make it clearer.