Add back button to practice window
Added a back button to the top right of the practice window, allowing the user to exit back to the setup screen at any time during practice to select a new song or change their input device.

Fixed the behaviour of the application when the user closes it, now properly waiting for the scoring process to finish before force exiting to prevent errors. This means the user can now safely exit the application during song playback.